+++ tests/include/test_globals.hpp
@@ -1,0 +1,6 @@
+#ifndef TEST_GLOBALS_H_
+#define TEST_GLOBALS_H_
+
+
+
+#endif // TEST_GLOBALS_H_